+++ Sources/TuistLoader/Models+ManifestMappers/TargetAction+ManifestMapper.swift
@@ -20,16 +20,28 @@ extension TuistCore.TargetAction {
         let outputFileListPaths = try manifest.outputFileListPaths.map { try generatorPaths.resolve(path: $0) }
         let basedOnDependencyAnalysis = manifest.basedOnDependencyAnalysis
         let path = try manifest.path.map { try generatorPaths.resolve(path: $0) }
-        return TargetAction(name: name,
-                            order: order,
-                            tool: tool,
-                            path: path,
-                            arguments: arguments,
-                            inputPaths: inputPaths,
-                            inputFileListPaths: inputFileListPaths,
-                            outputPaths: outputPaths,
-                            outputFileListPaths: outputFileListPaths,
-                            basedOnDependencyAnalysis: basedOnDependencyAnalysis)
+
+        if let embeddedScript = manifest.embeddedScript {
+            return TargetAction(name: name,
+                                order: order,
+                                script: embeddedScript,
+                                inputPaths: inputPaths,
+                                inputFileListPaths: inputFileListPaths,
+                                outputPaths: outputPaths,
+                                outputFileListPaths: outputFileListPaths,
+                                basedOnDependencyAnalysis: basedOnDependencyAnalysis)
+        } else {
+            return TargetAction(name: name,
+                                order: order,
+                                tool: tool,
+                                path: path,
+                                arguments: arguments,
+                                inputPaths: inputPaths,
+                                inputFileListPaths: inputFileListPaths,
+                                outputPaths: outputPaths,
+                                outputFileListPaths: outputFileListPaths,
+                                basedOnDependencyAnalysis: basedOnDependencyAnalysis)
+        }
     }
 }